pyproject.toml
[project]
name = "nlp-customer-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "hnswlib>=0.8.0",
    "pydantic>=2.11.3",
    "pyyaml>=6.0.2",
    "sentence-transformers>=4.0.2",
    "flair>=0.15.1",
    "mkdocs-material>=9.6.11",
    "rapidfuzz>=3.13.0",
    "streamlit>=1.44.1",
    "tqdm>=4.67.1",
    "textblob>=0.19.0",
    "streamlit-autorefresh>=1.0.1",
    "mlflow>=2.21.3",
    "prefect>=3.3.4",
    "zmq>=0.0.0",
    "loguru>=0.7.3",
]

[tool.ruff]
line-length = 190
